refactor; moved CDFToml out
