+++ cognite_toolkit/_cdf_tk/commands/build.py
@@ -21,6 +21,7 @@ from rich import print
 from rich.panel import Panel
 
 from cognite_toolkit._cdf_tk._parameters import ParameterSpecSet
+from cognite_toolkit._cdf_tk.cdf_toml import CDFToml
 from cognite_toolkit._cdf_tk.client import ToolkitClient
 from cognite_toolkit._cdf_tk.commands._base import ToolkitCommand
 from cognite_toolkit._cdf_tk.constants import (
@@ -31,7 +32,6 @@ from cognite_toolkit._cdf_tk.constants import (
 from cognite_toolkit._cdf_tk.data_classes import (
     BuildConfigYAML,
     BuildVariables,
-    CDFToml,
     ModuleDirectories,
 )
 from cognite_toolkit._cdf_tk.exceptions import (
