+++ cognite_toolkit/_cdf_tk/data_classes/__init__.py
@@ -1,5 +1,4 @@
 from ._build_variables import BuildVariable, BuildVariables
-from ._cdf_toml import CDFToml
 from ._config_yaml import (
     BuildConfigYAML,
     BuildEnvironment,
@@ -11,7 +10,6 @@ from ._config_yaml import (
 from ._module_directories import ModuleDirectories, ModuleLocation
 
 __all__ = [
-    "CDFToml",
     "InitConfigYAML",
     "ConfigYAMLs",
     "BuildConfigYAML",
